moving drstemplates.h contents into drstemplates.c
